Struct frustum_query::query::FrustumQuery

pub struct FrustumQuery { /* fields omitted */ }

FrustumQueryObject

A data structure with a collection of frustum query objects

Methods

impl FrustumQuery

fn new(objects: Vec<FrustumQueryObject>) -> FrustumQuery

Create a new FurstumQuery from a colletion of objects

fn execute(&self, frustum: &Frustum, result: &mut Vec<bool>)

Query if a list of objects are in the frustum